+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### BREAKING CHANGES
+- Dropped support for Node.js 14.
+
+### Added
+- Support for Node.js 20.
+
 ## [8.0.1] - 2022-07-30
 ### No changes
 - Updated devDependencies.