+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### BREAKING CHANGES
 - Dropped support for Node.js 14.
 - sortArrayByObjectProperty now throws a TypeError when it encounters elements of different types.
+- Convert package to ESM.
 
 ### Added
 - Support for Node.js 20.