+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [9.0.0] - 2023-06-14
 
 ### BREAKING CHANGES
 
@@ -176,7 +178,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Initial code.
 -   Setup for this project.
 
-[Unreleased]: https://github.com/Ionaru/array-utils/compare/8.0.1...HEAD
+[Unreleased]: https://github.com/Ionaru/array-utils/compare/9.0.0...HEAD
+[9.0.0]: https://github.com/Ionaru/array-utils/compare/8.0.1...9.0.0
 [8.0.1]: https://github.com/Ionaru/array-utils/compare/8.0.0...8.0.1
 [8.0.0]: https://github.com/Ionaru/array-utils/compare/7.0.0...8.0.0
 [7.0.0]: https://github.com/Ionaru/array-utils/compare/6.3.0...7.0.0